Improve timestamp/duration formatting and computing (ns and cycles)
Nanoseconds from origin (instead of from Epoch) are printed with digit
separators and the last nine digits are printed with a different color.
This is much more easy to read.

Cycles are printed with digit separators.

In the `Timestamp` object, when possible, nanoseconds from origin are
computed using the sum of the offset in cycles and the value in cycles,
so as to avoid two distinct nanosecond results (which are then summed
up) which can lead to a 1-ns error (comment in code explains the
reason).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>