Packet checkpoints: always create the last checkpoint, even on error
Jacques CTF needs to be able to inspect as many data as possible until a
potential decoding error. For packet checkpoints, this means that the
last checkpoint should be the last completely decodable event record.
Then at least we can inspect complete event records before and including
this last event record. Any remaining data can be inspected, but it
cannot be linked to an event record object.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>