Decouple data service from state
As of this patch, a `DataStreamFile` object offers packet index entries
as well as `Packet` objects. A `Packet` object offers event records and
data regions.

A `DataStreamFileState` owns a `DataStreamFile`, and also contains a
vector of packet states, one of which is active. A `PacketState` owns a
`Packet`, and also contains a current offset within said packet.

This will make it easier to reuse the `DataStreamFile` and `Packet`
objects in other projects without them being binded to the application's
state.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>