Make search threaded with animation with ERT ID/name query
Searching by ERT ID or name can be long because we need to decode each
event record, in order, to find the next occurence. This used to block
the UI with no "currently searching" indication.

When searching by ERT ID/name, animate the search input's border and
launch the search operation in a temporary thread.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>