Support bright colors
It looks like you can use bright colors (SGR codes 90 to 97, which
correspond to colors 8 to 15) using the color number directly in a color
pair. This is nice because "bold" becomes its own attribute, dissociated
from bright colors. This means Jacques CTF has access to more styling
possibilities for terminals which apply this distinction, like kitty.

However some terminals do not support support more than 8 colors and
this is indicated by the ncurses variable `COLORS`. In this case, fall
back to using the bold attribute to make the color bright (but also the
font bold).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>