Use "data stream file" terminology instead of "data stream"
There were a few places were "data stream" was used when it should
really be "data stream file". Jacques CTF does not work at the data
stream level, in that it won't merge data stream files into a single
data stream view, they are all treated independently. Not that it would
not be a nice feature to have.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>