Packet data view: highlight offsets of all rows containing current region
Instead of highlighting the offset at the left of the row containing the
first bit of the current packet region, highlight the offsets of all the
rows containing any bit of the current packet region.

Visually, you can still see the information we had previously by looking
at the top highlighted offset.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>