Table view: handle cell style outside _drawCell()
TableView::_drawCells() handles setting the current style (using the
stylist) of the cell now. It's easier to put this logic outside
TableView::_drawCell(). This allows having a single cell having a
warning or error style even if it's selected.

All styles are per-cell now. TableView::_drawCells() does not check the
first cell's style to apply it to the whole row anymore, so if a
specialized table view needs to make the whole row in warning/error
style, it needs to set each cell in this style.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>