Data stream file table view: time cells: use error style when incorrect
In the data stream file table view, when, for a given row, the end time
is less than the beginning time, make those cells have the error state.

LTTng has bugs which can lead to this situation.

A whole row's cells were in the error state when, for this data stream
file, there was an error (typically found during the packet indexing
operation). Now, to distinguish the time errors from the file's error,
only the path cell is in the error state when the file has an error.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>