Add `list-packets` command
This new command lists the packets of a given data stream file and their
details. It's pretty much the same information that you find in the
Packets screen of the text UI.

Currently you must pass the `--machine` option because I did not
implement the human-readable format yet (which would be the default).
When I do, `--machine` will become optional.

The machine format is a simple list of CSV. It's safe to use CSV here
because there's no text from the trace data, only numbers and boolean
properties. When a property is missing, the value is empty.

I chose CSV to make it easy to write a shell script which can pick
individual values from the output with `cut` and other readily available
tools. For example you can isolate the 10th packet like this:

    $ jacquesctf list-packets -m /path/to/stream/file > packets.csv
    $ line=$(sed '10q;d' packets.csv)
    $ offset=$(echo "$line" | cut -d, -f2)
    $ size=$(echo "$line" | cut -d, -f3)
    $ dd if=/path/to/stream/file of=my-packet ibs=1 \
         skip=$offset count=$size

You can print a textual header with `--header`. This will look nice once
you open the resulting output with a spreadsheet.

When the data stream file is empty, the command prints nothing, not
even the header.

Note that the last field of a row is `yes` or `no`, whether or not the
packet is valid. The packet is only validated up to the beginning of the
first event record. An invalid packet could have invalid total and/or
content sizes.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>