Make building the `inspect` command optional
This patch makes building the `inspect` command optional.

The build system only builds the `inspect` if (both):

* The `OPT_WITHOUT_INSPECT_CMD` option is off.
* `find_package (Curses)` finds the dependency.

When the `inspect` command isn't built, `jacquesctf inspect ...` or
simply `jacquesctf ...` indicates that the command could exist, but it's
not available in this build.

My goal is creating a Qt-based GUI version of Jacques CTF (new command),
which will also be optional at build time, so it makes sense to also
make the TUI inspection command optional.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>