Make `jacques::TableView` inherit `jacques::CommonInspectTableView`
One important change with this patch is that a `TableView` doesn't ask
the object if an index exists (_hasIndex()) anymore, but asks its total
row count (_rowCount()) instead.

When the total row count of a concrete table view changes, it calls
TableView::_updateCounts(), which itself calls
CommonInspectTableView::_updateCounts() with the results of the virtual
TableView::_rowCount().

No functional change intended.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>