Make application state classes common
Application state classes are now part of `jacquesctf/inspect-common`.

This makes it possible to reuse state objects in other inspection UIs.

The `State` class is renamed to `AppState` (common version) and
`InspectCmdState` (specific to the `inspect` command).

`AppState` is almost like `State` used to be, but it has no observers,
calling the virtual _activeDsFileChanged(), _activePktChanged(), and
_curOffsetInPktChanged() instead so that the implementation may use its
own signalling strategy. `InspectCmdState` uses observers as before.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>