AppState: rename _activeDsFileChanged() -> _activeDsFileAndPktChanged()
This patch makes an `AppState` call _activeDsFileAndPktChanged() when
the data stream file changes.

The main difference is that, without this patch, calling
AppState::gotoDsFile() makes the object call _activeDsFileChanged()
first and then _activePktChanged(), whereas with this patch it only
calls _activeDsFileAndPktChanged(). This is because when the active data
stream file changes, the active packet must change too.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>